Replace speech synthesis with audio playback
On Firefox on Linux, the speech synthesis did not work.

Additionally, on all platforms, it sounded wonky and slightly wrong,
which is not the kind of quality we expect from such pristine
productions.

Finally, using MP3 gives a more consistent output across platforms,
without adding all that much loading time considering the benefit.